ci-ng/requirements.txt
Django==1.9.4
django-annoying==0.9.0
django-debug-toolbar==1.4
django-npm==0.1.4
sqlparse==0.1.19
